third_party/blink/web_tests/external/wpt/css/css-backgrounds/parsing/background-position-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-position value 'center right 7%'
  assert_equals: expected "93% 50%" but got "right 7% 50%"
[FAIL] Property background-position value 'center top 8px'
  assert_equals: expected "50% 8px" but got "50% top 8px"
[FAIL] Property background-position value 'left 10px center'
  assert_equals: expected "10px 50%" but got "left 10px 50%"
[FAIL] Property background-position value 'right 11% bottom'
  assert_equals: expected "89% 100%" but got "right 11% 100%"
[FAIL] Property background-position value 'left 12px top 13px'
  assert_equals: expected "12px 13px" but got "left 12px top 13px"
[FAIL] Property background-position value 'right top 14%'
  assert_equals: expected "100% 14%" but got "100% top 14%"
[FAIL] Property background-position value 'top 15px center'
  assert_equals: expected "50% 15px" but got "50% top 15px"
[FAIL] Property background-position value 'bottom 16% left'
  assert_equals: expected "0% 84%" but got "0% bottom 16%"
[FAIL] Property background-position value 'top 17px right -18px'
  assert_equals: expected "calc(100% + 18px) 17px" but got "right -18px top 17px"
[FAIL] Property background-position value 'bottom right 19%'
  assert_equals: expected "81% 100%" but got "right 19% 100%"
Harness: the test ran to completion.
